POC Completed:  Generate Images from Open AI
Portfolio - Demo project 2

In order to develop a documentation website we wanted to use a specific public API. We have noticed that this API didn't return any images which is an issue for our demo project.

This is why we have implemented this POC using the OpenAi module to confirm whether it could generate the image documentation we would need for our documentation website project.

We set up the openai package to generate an image with a specific prompt.
However the prompt is very specific as we want special data from the API and we don't have the expected result (ie. configuration.js)

Therefore we are not sure that we are going to use this solution or this API for our Documentation Demo Project.

We might have a look at the Wikipedia API to check the feasibility of our project.